feat: add sweet alert 2 and style: style apparel form

+++ src/pages/WardrobePage/WardrobePage.jsx
@@ -3,11 +3,13 @@ import {
   deleteApparelService,
   getAllApparelService,
   getUniqueCategories,
+  swalBasicSettings,
 } from "../../utilities/wardrobe-service";
 import { LuPlusSquare } from "react-icons/lu";
 import debug from "debug";
 import { Link } from "react-router-dom";
 import ApparelRow from "../../components/ApparelRow/ApparelRow";
+import Swal from "sweetalert2";
 
 const log = debug("nextfit:src:pages:WardrobePage");
 
@@ -26,20 +28,35 @@ function WardrobePage() {
   const categories = getUniqueCategories(apparel);
 
   const handleDelete = async (apparelID) => {
-    try {
-      await deleteApparelService(apparelID);
-      const remainingApparel = apparel.filter((item) => item._id !== apparelID);
-      log("deleted apparel:", remainingApparel);
-      setApparel(remainingApparel);
-    } catch (err) {
-      console.error(err);
+    const prompt = await Swal.fire({
+      ...swalBasicSettings("Proceed to delete?", "warning"),
+      showCancelButton: true,
+      confirmButtonText: "DELETE",
+      cancelButtonText: "CANCEL",
+    });
+
+    if (prompt.isConfirmed) {
+      try {
+        await deleteApparelService(apparelID);
+        const remainingApparel = apparel.filter(
+          (item) => item._id !== apparelID
+        );
+        log("deleted apparel:", remainingApparel);
+        setApparel(remainingApparel);
+        Swal.fire(swalBasicSettings("Deleted!", "success"));
+      } catch (err) {
+        console.error(err);
+        Swal.fire({
+          ...swalBasicSettings("Error", "error"),
+          text: "Something went wrong",
+        });
+      }
     }
   };
 
   return (
     <div className="mx-auto max-w-screen-xl p-4">
-      <header className="flex mx-4 mb-4 font-bebas text-3xl tracking-wider justify-between">
-        <h1>My Wardrobe</h1>
+      <header className="flex mx-4 mb-4 font-bebas text-3xl tracking-wider justify-end">
         <Link to="/wardrobe/new" className="flex items-center justify-center">
           <LuPlusSquare className="text-4xl" />
           <span className="font-inter font-semibold tracking-normal text-xl">
